+++ src/lib/format_file.ts
@@ -3,11 +3,16 @@ import {extname} from 'node:path';
 
 import {load_package_json} from './package_json.js';
 
-export const format_file = async (id: string, content: string): Promise<string> => {
+// TODO improve API - is `id` badly duplcating `filepath`?
+export const format_file = async (
+	id: string,
+	content: string,
+	options?: prettier.Options | undefined,
+): Promise<string> => {
 	const parser = infer_parser(id);
 	if (!parser) return content;
 	const config: any = (await load_package_json()).prettier;
-	return prettier.format(content, {...config, parser});
+	return prettier.format(content, {...config, parser, ...options});
 };
 
 // This is a lot faster than `prettier.getFileInfo`