+++ src/lib/package_json.ts
@@ -169,7 +169,13 @@ export const to_package_exports = (paths: string[]): PackageJsonExports => {
 		.sort((a, b) => (a === 'index.ts' ? -1 : b === 'index.ts' ? 1 : a.localeCompare(b)));
 	const exports: PackageJsonExports = {};
 	for (const path of sorted) {
-		if (path.endsWith('.ts')) {
+		if (path.endsWith('.json.d.ts')) {
+			const json_path = path.substring(0, path.length - 5);
+			exports['./' + json_path] = {
+				default: IMPORT_PREFIX + json_path, // assuming a matching json file
+				types: IMPORT_PREFIX + path,
+			};
+		} else if (path.endsWith('.ts') && !path.endsWith('.d.ts')) {
 			const js_path = replace_extension(path, '.js');
 			const key = path === 'index.ts' ? '.' : './' + js_path;
 			exports[key] = {