+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @grogarden/gro
 
+## 0.94.2
+
+### Patch Changes
+
+- 9c860a5: support prettier options to format_file
+
+## 0.94.1
+
+### Patch Changes
+
+- c18723e: support typed json exports
+
 ## 0.94.0
 
 ### Minor Changes